+++ torchmetrics/text/wer.py
@@ -65,7 +65,8 @@ class WER(Metric):
         >>> metric(predictions, references)
         tensor(0.5000)
     """
-
+    is_differentiable = False
+    higher_is_better = False
     error: Tensor
     total: Tensor